requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
billiard==3.3.0.23
boto3==1.4.4
botocore==1.5.37
celery==3.1.25
Django==1.10.6
django-celery==3.2.1
django-debug-toolbar==1.7
djangorestframework==3.6.2
docutils==0.13.1
futures==3.0.5
gunicorn==19.7.1
jmespath==0.9.2
kombu==3.0.37
olefile==0.44
packaging==16.8
Pillow
psycopg2
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
redis==2.10.5
s3transfer==0.1.10
six==1.10.0
sqlparse
